+++ modules/processing_args.py
@@ -65,7 +65,7 @@ def task_specific_kwargs(p, model):
     elif (sd_models.get_diffusers_task(model) == sd_models.DiffusersTaskType.INPAINTING or is_img2img_model) and len(getattr(p, 'init_images', [])) > 0:
         if shared.sd_model_type == 'sdxl' and hasattr(model, 'register_to_config'):
             model.register_to_config(requires_aesthetics_score = False)
-        if p.detailer:
+        if p.detailer_enabled:
             p.ops.append('detailer')
         else:
             p.ops.append('inpaint')